refactor default, forum
refactor default, forum

done refactor api
